requirements.txt
flask
openai
langchain
langchain-community
python-dotenv